src/ai-consumer/pyproject.toml
[project]
name = "ai-consumer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aio-pika>=9.5.8",
    "google-genai>=1.56.0",
    "pika>=1.3.2",
    "redis>=7.1.0",
    "sqlalchemy>=2.0.45",
]
